# backend/app/swagger.yaml
swagger: "2.0"
info:
  title: School Management System API
  description: |
    REST API for the School Management System (SMS).

    ## Authentication
    Most endpoints require a JWT Bearer token. Log in via **POST /auth/login** to obtain
    `access_token` and `refresh_token`. Pass the access token as:
    ```
    Authorization: Bearer <access_token>
    ```
    The access token expires in **15 minutes**. Use **POST /auth/refresh** with the refresh token
    to get a new access token (refresh tokens expire in 7 days).

    ## Roles
    | Role | Access |
    |------|--------|
    | `admin` | All endpoints |
    | `teacher` | Students, attendance, grades, leave review |
    | `student` | Own data only |
    | `parent` | Parent portal, leave submissions, notifications |

    ## Response Envelope
    All responses follow:
    ```json
    { "success": true, "data": {}, "message": "...", "errors": null }
    ```

    ## Dev Seed Credentials
    | Role | Email | Password |
    |------|-------|----------|
    | Admin | admin@school.sms | Admin@1234 |
    | Teacher | priya.sharma@school.sms | Teacher@123 |
    | Student | alice.johnson@school.sms | Student@123 |
    | Parent | robert.johnson@parents.school.sms | Parent@123 |

  version: "1.0.0"
  contact:
    email: admin@school.sms
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Format: Bearer <JWT access token>"
...
definitions:
...
  Student:
    type: object
    properties:
      id:
        type: integer
        example: 1
      admission_no:
        type: string
        example: ADM2024001
      first_name:
        type: string
        example: Alice
      last_name:
        type: string
        example: Johnson
      full_name:
        type: string
        example: Alice Johnson
      date_of_birth:
        type: string
        format: date
        example: "2012-03-15"
      gender:
        type: string
        enum: [Male, Female, Other]
        example: Female
      admission_date:
        type: string
        format: date
        example: "2024-06-01"
      blood_group:
        type: string
        example: A+
        nullable: true
      address:
        type: string
        nullable: true
      phone:
        type: string
        nullable: true
      is_active:
        type: boolean
        example: true
  LeaveApplication:
    type: object
    properties:
      id:
        type: integer
        example: 1
      student_id:
        type: integer
        example: 1
      parent_id:
        type: integer
        example: 1
      from_date:
        type: string
        format: date
        example: "2026-06-10"
      to_date:
        type: string
        format: date
        example: "2026-06-10"
      duration_days:
        type: integer
        example: 1
      reason:
        type: string
        example: Doctor appointment for annual check-up
      leave_type:
        type: string
        enum: [sick, family, personal, other]
        example: sick
      status:
        type: string
        enum: [pending, approved, rejected]
        example: pending
      reviewer_remarks:
        type: string
        nullable: true
      reviewed_at:
        type: string
        format: date-time
        nullable: true
      created_at:
        type: string
        format: date-time
  Notification:
    type: object
    properties:
      id:
        type: integer
        example: 1
      type:
        type: string
        enum: [absence, low_marks, fee_due, message, announcement, leave_update]
        example: announcement
      title:
        type: string
        example: School Annual Day 2024
      body:
        type: string
        example: Annual Day function scheduled on 25th December.
      reference_id:
        type: integer
        nullable: true
      reference_type:
        type: string
        nullable: true
      is_read:
        type: boolean
        example: false
      created_at:
        type: string
        format: date-time

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /students:
    get:
      tags: [Students]
      summary: List students
      description: Paginated, searchable list. **Admin or Teacher only.**
      security:
        - Bearer: []
      parameters:
        - in: query
          name: page
          type: integer
          default: 1
        - in: query
          name: per_page
          type: integer
          default: 20
        - in: query
          name: search
          type: string
          description: Search by name or admission number
      responses:
        200:
          description: List of students
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/definitions/Student'
        401:
          description: Unauthorized
        403:
          description: Insufficient permissions

    post:
      tags: [Students]
      summary: Enroll a new student
      description: Creates a full student profile. **Admin only.**
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [admission_no, first_name, last_name, date_of_birth, gender, admission_date, user_id]
            properties:
              user_id:
                type: integer
                description: ID of the existing user with role=student
                example: 4
              admission_no:
                type: string
                example: ADM2024001
              first_name:
                type: string
                example: Alice
              last_name:
                type: string
                example: Johnson
              date_of_birth:
                type: string
                format: date
                example: "2012-03-15"
              gender:
                type: string
                enum: [Male, Female, Other]
                example: Female
              admission_date:
                type: string
                format: date
                example: "2024-06-01"
              blood_group:
                type: string
                example: A+
              address:
                type: string
              phone:
                type: string
      responses:
        201:
          description: Student enrolled
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Student'
        400:
          description: Invalid input
        401:
          description: Unauthorized
        403:
          description: Admin access required
        409:
          description: Admission number already exists
  /students/{student_id}:
    get:
      tags: [Students]
      summary: Get student by ID
      security:
        - Bearer: []
      parameters:
        - in: path
          name: student_id
          type: integer
          required: true
          example: 1
      responses:
        200:
          description: Student found
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Student'
        404:
          description: Student not found

    put:
      tags: [Students]
      summary: Update student profile
      description: Update editable student fields. **Admin only.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: student_id
          type: integer
          required: true
          example: 1
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              first_name:
                type: string
              last_name:
                type: string
              address:
                type: string
              phone:
                type: string
              blood_group:
                type: string
      responses:
        200:
          description: Student updated
        404:
          description: Student not found

    delete:
      tags: [Students]
      summary: Deactivate student (soft delete)
      description: Sets `is_active=false`. **Admin only.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: student_id
          type: integer
          required: true
          example: 1
      responses:
        200:
          description: Student deactivated
        404:
          description: Student not found

  # ── Parent Portal ─────────────────────────────────────────────────────────────
  /parent-portal/dashboard:
    get:
      tags: [Parent Portal]
      summary: Parent dashboard
      description: |
        Overview of all children linked to the authenticated parent.
        Returns summary cards for each child (attendance %, latest fee status, pending leaves).
        **Parent role only.**
      security:
        - Bearer: []
      responses:
        200:
          description: Dashboard data
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  children:
                    type: array
                    items:
                      type: object
        401:
          description: Unauthorized
        403:
          description: Parent access required or parent profile not linked
  /parent-portal/children:
    get:
      tags: [Parent Portal]
      summary: List linked children
      description: Returns all students linked to the authenticated parent. **Parent only.**
      security:
        - Bearer: []
      responses:
        200:
          description: Children list
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  children:
                    type: array
                    items:
                      $ref: '#/definitions/Student'
        403:
          description: Access denied
  /parent-portal/children/{child_id}/attendance:
    get:
      tags: [Parent Portal]
      summary: Child attendance
      description: Monthly attendance calendar for a specific child. **Parent only.** 403 if not your child.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: child_id
          type: integer
          required: true
          example: 1
        - in: query
          name: month
          type: integer
          example: 6
          description: Month (1-12). Defaults to current month.
        - in: query
          name: year
          type: integer
          example: 2026
          description: Year. Defaults to current year.
      responses:
        200:
          description: Attendance records
        403:
          description: Not your child
  /parent-portal/children/{child_id}/grades:
    get:
      tags: [Parent Portal]
      summary: Child grades
      description: Exam results and report card for a specific child. **Parent only.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: child_id
          type: integer
          required: true
          example: 1
      responses:
        200:
          description: Grade records
        403:
          description: Not your child
  /parent-portal/children/{child_id}/fees:
    get:
      tags: [Parent Portal]
      summary: Child fee status
      description: Fee records and payment history for a specific child. **Parent only.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: child_id
          type: integer
          required: true
          example: 1
      responses:
        200:
          description: Fee records
        403:
          description: Not your child

  # ── Leave Applications ────────────────────────────────────────────────────────
  /leave-applications:
    get:
      tags: [Leave Applications]
      summary: List leave applications
      description: Returns all leave applications submitted by the authenticated parent. **Parent only.**
      security:
        - Bearer: []
      responses:
        200:
          description: Leave applications
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  leaves:
                    type: array
                    items:
                      $ref: '#/definitions/LeaveApplication'
        403:
          description: Parent access required

    post:
      tags: [Leave Applications]
      summary: Submit leave application
      description: |
        Parent submits a leave request for their child. The child must be linked to the parent.
        `from_date` cannot be in the past. **Parent only.**
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [student_id, from_date, to_date, leave_type, reason]
            properties:
              student_id:
                type: integer
                description: Must be one of your linked children
                example: 1
              from_date:
                type: string
                format: date
                example: "2026-06-15"
              to_date:
                type: string
                format: date
                example: "2026-06-15"
              leave_type:
                type: string
                enum: [sick, family, personal, other]
                example: sick
              reason:
                type: string
                example: Child has fever and requires rest
      responses:
        201:
          description: Leave application submitted
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/LeaveApplication'
        400:
          description: Missing fields or date in the past
        403:
          description: Not your child
  /leave-applications/{leave_id}/review:
    put:
      tags: [Leave Applications]
      summary: Approve or reject a leave application
      description: Admin or Teacher reviews a pending leave application. **Admin or Teacher only.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: leave_id
          type: integer
          required: true
          example: 1
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [status]
            properties:
              status:
                type: string
                enum: [approved, rejected]
                example: approved
              remarks:
                type: string
                example: Approved. Please submit medical certificate on return.
      responses:
        200:
          description: Leave reviewed
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/LeaveApplication'
        400:
          description: Invalid status
        403:
          description: Admin or Teacher access required
        404:
          description: Leave application not found

  # ── Notifications ─────────────────────────────────────────────────────────────
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      description: Returns notifications for the authenticated user. Supports filtering by unread. **All roles.**
      security:
        - Bearer: []
      parameters:
        - in: query
          name: unread
          type: boolean
          default: false
          description: If true, return only unread notifications
      responses:
        200:
          description: Notifications list
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/definitions/Notification'
        401:
          description: Unauthorized
  /notifications/{notification_id}/read:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      description: Marks a specific notification as read. Only the notification owner can mark it. **All roles.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: notification_id
          type: integer
          required: true
          example: 1
      responses:
        200:
          description: Notification marked as read
        404:
          description: Notification not found or not yours